+++ store/pruning.go
@@ -1,18 +1,15 @@
 package store
 
-import "fmt"
+import (
+	"fmt"
+)
 
 // PruneBlocks removes block up to (but not including) a height. It returns number of blocks pruned.
-func (s *DefaultStore) PruneBlocks(heightInt int64) (uint64, error) {
-	if heightInt <= 0 {
-		return 0, fmt.Errorf("height must be greater than 0")
+func (s *DefaultStore) PruneBlocks(base, height uint64) (uint64, error) {
+	if base <= 0 {
+		return 0, fmt.Errorf("from height must be greater than 0")
 	}
 
-	height := uint64(heightInt)
-	if height > s.Height() {
-		return 0, fmt.Errorf("cannot prune beyond the latest height %v", s.height)
-	}
-	base := s.Base()
 	if height < base {
 		return 0, fmt.Errorf("cannot prune to height %v, it is lower than base height %v",
 			height, base)
@@ -22,13 +19,10 @@ func (s *DefaultStore) PruneBlocks(heightInt int64) (uint64, error) {
 	batch := s.db.NewBatch()
 	defer batch.Discard()
 
-	flush := func(batch Batch, base uint64) error {
+	flush := func(batch Batch, height uint64) error {
 		err := batch.Commit()
 		if err != nil {
-			return fmt.Errorf("prune up to height %v: %w", base, err)
-		}
-		if ok := s.SetBase(base); !ok {
-			return fmt.Errorf("set base height: %v", base)
+			return fmt.Errorf("flush batch to disk: height %d: %w", height, err)
 		}
 		return nil
 	}
