+++ store/storeIface.go
@@ -1,31 +1,44 @@
 package store
 
 import (
+	"github.com/dymensionxyz/dymint/types"
 	tmstate "github.com/tendermint/tendermint/proto/tendermint/state"
 	tmtypes "github.com/tendermint/tendermint/types"
-
-	"github.com/dymensionxyz/dymint/types"
 )
+
+// KVStore encapsulates key-value store abstraction, in minimalistic interface.
+//
+// KVStore MUST be thread safe.
+type KVStore interface {
+	Get(key []byte) ([]byte, error)        // Get gets the value for a key.
+	Set(key []byte, value []byte) error    // Set updates the value for a key.
+	Delete(key []byte) error               // Delete deletes a key.
+	NewBatch() Batch                       // NewBatch creates a new batch.
+	PrefixIterator(prefix []byte) Iterator // PrefixIterator creates iterator to traverse given prefix.
+}
+
+// Batch enables batching of transactions.
+type Batch interface {
+	Set(key, value []byte) error // Accumulates KV entries in a transaction.
+	Delete(key []byte) error     // Deletes the given key.
+	Commit() error               // Commits the transaction.
+	Discard()                    // Discards the transaction.
+}
+
+// Iterator enables traversal over a given prefix.
+type Iterator interface {
+	Valid() bool
+	Next()
+	Key() []byte
+	Value() []byte
+	Error() error
+	Discard()
+}
 
 // Store is minimal interface for storing and retrieving blocks, commits and state.
 type Store interface {
-	// NewBatch creates a new db batch.
+	// NewStoreBatch creates a new db batch.
 	NewBatch() Batch
-
-	// Height returns height of the highest block in store.
-	Height() uint64
-
-	// NextHeight returns the next height that expected to be stored in store.
-	NextHeight() uint64
-
-	// SetHeight sets the height saved in the Store if it is higher than the existing height.
-	SetHeight(height uint64) bool
-
-	// Base returns height of the lowest block in store.
-	Base() uint64
-
-	// SetBase sets the height saved in the Store for the lowest block
-	SetBase(height uint64) bool
 
 	// SaveBlock saves block along with its seen commit (which will be included in the next block).
 	SaveBlock(block *types.Block, commit *types.Commit, batch Batch) (Batch, error)
@@ -58,5 +71,9 @@ type Store interface {
 	LoadValidators(height uint64) (*tmtypes.ValidatorSet, error)
 
 	// Pruning functions
-	PruneBlocks(height int64) (uint64, error)
+	PruneBlocks(from, to uint64) (uint64, error)
+
+	// Left those for backward comptability. Should be removed in the future.
+	SetHeight(height uint64)
+	Height() uint64
 }
